+++ sql/postgres/migrations/20250820083141_update_two_factor_columns.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE owners DROP COLUMN IF EXISTS otp_data;